+++ src/main/java/narcibot/IncorrectFormatException.java
@@ -1,3 +1,4 @@
+package narcibot;
 public class IncorrectFormatException extends Exception {
     public IncorrectFormatException (String errorMessage) {
         super(errorMessage);
