+++ src/main/java/narcibot/Narcibot.java
@@ -1,4 +1,4 @@
-
+package narcibot;
 import java.io.*;
 import java.util.ArrayList;
 import java.util.Scanner;
