+++ src/main/java/narcibot/Task.java
@@ -1,3 +1,4 @@
+package narcibot;
 public class Task {
     private String name;
     private boolean done;
